store: Move creating a layout and starting a subgraph out of entities
